Update ShoppingBill
This Java program will ask the number of products bought and the prices. It will count the total price and will automatically add VAT. It'll ask the user to pay. If the user enter less amount, the program will continue to ask the remain amount. If the user enters the amount more than the bill, it'll count the change.